+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -1,35 +1,11 @@
 package kitchenpos.application;
 
 import java.util.List;
-import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.ui.dto.request.MenuGroupCreateRequest;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
 
-@Service
-public class MenuGroupService {
-    private final MenuGroupDao menuGroupDao;
+public interface MenuGroupService {
+    MenuGroup create(MenuGroupCreateRequest request);
 
-    public MenuGroupService(final MenuGroupDao menuGroupDao) {
-        this.menuGroupDao = menuGroupDao;
-    }
-
-    @Transactional
-    public MenuGroup create(final MenuGroupCreateRequest request) {
-        final var requestedName = request.getName();
-        validateName(requestedName);
-        return menuGroupDao.save(new MenuGroup(requestedName));
-    }
-
-    private void validateName(final String name) {
-        if (!StringUtils.hasText(name)) {
-            throw new IllegalArgumentException("유효하지 않은 메뉴 그룹명 : " + name);
-        }
-    }
-
-    public List<MenuGroup> list() {
-        return menuGroupDao.findAll();
-    }
+    List<MenuGroup> list();
 }
