+++ src/main/java/kitchenpos/application/old/JdbcMenuService.java
@@ -1,0 +1,108 @@
+package kitchenpos.application.old;
+
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import kitchenpos.application.MenuService;
+import kitchenpos.dao.MenuDao;
+import kitchenpos.dao.MenuGroupDao;
+import kitchenpos.dao.MenuProductDao;
+import kitchenpos.dao.ProductDao;
+import kitchenpos.domain.Menu;
+import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.Product;
+import kitchenpos.ui.dto.request.MenuCreateRequest;
+import kitchenpos.ui.dto.request.MenuProductRequest;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Transactional(readOnly = true)
+@Service
+public class JdbcMenuService implements MenuService {
+    private final MenuDao menuDao;
+    private final MenuGroupDao menuGroupDao;
+    private final MenuProductDao menuProductDao;
+    private final ProductDao productDao;
+
+    public JdbcMenuService(
+            final MenuDao menuDao,
+            final MenuGroupDao menuGroupDao,
+            final MenuProductDao menuProductDao,
+            final ProductDao productDao
+    ) {
+        this.menuDao = menuDao;
+        this.menuGroupDao = menuGroupDao;
+        this.menuProductDao = menuProductDao;
+        this.productDao = productDao;
+    }
+
+    @Transactional
+    @Override
+    public Menu create(final MenuCreateRequest request) {
+        validateMenuProducts(request.getMenuProducts());
+
+        final var menuProducts = request.getMenuProducts()
+                .stream()
+                .map(req -> new MenuProduct(req.getProductId(), req.getQuantity()))
+                .collect(Collectors.toList());
+
+        final var menu = new Menu(
+                request.getName(),
+                request.getPrice(),
+                request.getMenuGroupId(),
+                menuProducts
+        );
+
+        final BigDecimal price = menu.getPrice();
+
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException();
+        }
+
+        if (!menuGroupDao.existsById(menu.getMenuGroupId())) {
+            throw new IllegalArgumentException();
+        }
+
+        BigDecimal sum = BigDecimal.ZERO;
+        for (final MenuProduct menuProduct : menuProducts) {
+            final Product product = productDao.findById(menuProduct.getProductId())
+                    .orElseThrow(IllegalArgumentException::new);
+            sum = sum.add(product.getPrice().multiply(BigDecimal.valueOf(menuProduct.getQuantity())));
+        }
+
+        if (price.compareTo(sum) > 0) {
+            throw new IllegalArgumentException();
+        }
+
+        final Menu savedMenu = menuDao.save(menu);
+
+        final Long menuId = savedMenu.getId();
+        final List<MenuProduct> savedMenuProducts = new ArrayList<>();
+        for (final MenuProduct menuProduct : menuProducts) {
+            menuProduct.setMenuId(menuId);
+            savedMenuProducts.add(menuProductDao.save(menuProduct));
+        }
+        savedMenu.setMenuProducts(savedMenuProducts);
+
+        return savedMenu;
+    }
+
+    private void validateMenuProducts(final List<MenuProductRequest> menuProducts) {
+        if (Objects.isNull(menuProducts) || menuProducts.isEmpty()) {
+            throw new IllegalArgumentException("유효하지 않은 메뉴 프로덕트 : " + menuProducts);
+        }
+    }
+
+    @Override
+    public List<Menu> list() {
+        final List<Menu> menus = menuDao.findAll();
+
+        for (final Menu menu : menus) {
+            menu.setMenuProducts(menuProductDao.findAllByMenuId(menu.getId()));
+        }
+
+        return menus;
+    }
+}
