+++ src/main/java/kitchenpos/application/old/JdbcOrderService.java
@@ -1,0 +1,121 @@
+package kitchenpos.application.old;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import kitchenpos.application.OrderService;
+import kitchenpos.dao.MenuDao;
+import kitchenpos.dao.OrderDao;
+import kitchenpos.dao.OrderLineItemDao;
+import kitchenpos.dao.OrderTableDao;
+import kitchenpos.domain.Order;
+import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.OrderTable;
+import kitchenpos.ui.dto.request.OrderChangeStatusRequest;
+import kitchenpos.ui.dto.request.OrderCreateRequest;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
+
+@Transactional(readOnly = true)
+@Service
+public class JdbcOrderService implements OrderService {
+    private final MenuDao menuDao;
+    private final OrderDao orderDao;
+    private final OrderLineItemDao orderLineItemDao;
+    private final OrderTableDao orderTableDao;
+
+    public JdbcOrderService(
+            final MenuDao menuDao,
+            final OrderDao orderDao,
+            final OrderLineItemDao orderLineItemDao,
+            final OrderTableDao orderTableDao
+    ) {
+        this.menuDao = menuDao;
+        this.orderDao = orderDao;
+        this.orderLineItemDao = orderLineItemDao;
+        this.orderTableDao = orderTableDao;
+    }
+
+    @Transactional
+    @Override
+    public Order create(final OrderCreateRequest request) {
+        final var orderLineItemRequests = request.getOrderLineItems();
+        if (CollectionUtils.isEmpty(orderLineItemRequests)) {
+            throw new IllegalArgumentException();
+        }
+
+        final var orderLineItems = orderLineItemRequests.stream()
+                .map(req -> new OrderLineItem(req.getMenuId(), req.getQuantity()))
+                .collect(Collectors.toList());
+
+        final var order = new Order(request.getOrderTableId(), orderLineItems);
+
+        final List<Long> menuIds = orderLineItems.stream()
+                .map(OrderLineItem::getMenuId)
+                .collect(Collectors.toList());
+
+        if (orderLineItems.size() != menuDao.countByIdIn(menuIds)) {
+            throw new IllegalArgumentException();
+        }
+
+        order.setId(null);
+
+        final OrderTable orderTable = orderTableDao.findById(order.getOrderTableId())
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (orderTable.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+
+        order.setOrderTableId(orderTable.getId());
+        order.setOrderStatus(OrderStatus.COOKING.name());
+        order.setOrderedTime(LocalDateTime.now());
+
+        final Order savedOrder = orderDao.save(order);
+
+        final Long orderId = savedOrder.getId();
+        final List<OrderLineItem> savedOrderLineItems = new ArrayList<>();
+        for (final OrderLineItem orderLineItem : orderLineItems) {
+            orderLineItem.setOrderId(orderId);
+            savedOrderLineItems.add(orderLineItemDao.save(orderLineItem));
+        }
+        savedOrder.setOrderLineItems(savedOrderLineItems);
+
+        return savedOrder;
+    }
+
+    @Override
+    public List<Order> list() {
+        final List<Order> orders = orderDao.findAll();
+
+        for (final Order order : orders) {
+            order.setOrderLineItems(orderLineItemDao.findAllByOrderId(order.getId()));
+        }
+
+        return orders;
+    }
+
+    @Transactional
+    @Override
+    public Order changeOrderStatus(final Long orderId, final OrderChangeStatusRequest request) {
+        final Order savedOrder = orderDao.findById(orderId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        if (Objects.equals(OrderStatus.COMPLETION.name(), savedOrder.getOrderStatus())) {
+            throw new IllegalArgumentException();
+        }
+
+        final OrderStatus orderStatus = OrderStatus.valueOf(request.getOrderStatus());
+        savedOrder.setOrderStatus(orderStatus.name());
+
+        orderDao.save(savedOrder);
+
+        savedOrder.setOrderLineItems(orderLineItemDao.findAllByOrderId(orderId));
+
+        return savedOrder;
+    }
+}
