+++ src/main/java/kitchenpos/application/old/JdbcTableGroupService.java
@@ -1,0 +1,97 @@
+package kitchenpos.application.old;
+
+import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import kitchenpos.application.TableGroupService;
+import kitchenpos.dao.OrderDao;
+import kitchenpos.dao.OrderTableDao;
+import kitchenpos.dao.TableGroupDao;
+import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.TableGroup;
+import kitchenpos.ui.dto.request.TableGroupCreateRequest;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
+
+@Transactional(readOnly = true)
+@Service
+public class JdbcTableGroupService implements TableGroupService {
+    private final OrderDao orderDao;
+    private final OrderTableDao orderTableDao;
+    private final TableGroupDao tableGroupDao;
+
+    public JdbcTableGroupService(final OrderDao orderDao, final OrderTableDao orderTableDao,
+                                 final TableGroupDao tableGroupDao) {
+        this.orderDao = orderDao;
+        this.orderTableDao = orderTableDao;
+        this.tableGroupDao = tableGroupDao;
+    }
+
+    @Transactional
+    @Override
+    public TableGroup create(final TableGroupCreateRequest request) {
+        final var orderTableIds = request.getOrderTables();
+
+        if (CollectionUtils.isEmpty(orderTableIds) || orderTableIds.size() < 2) {
+            throw new IllegalArgumentException();
+        }
+
+        final List<OrderTable> savedOrderTables = orderTableDao.findAllByIdIn(orderTableIds);
+
+        if (orderTableIds.size() != savedOrderTables.size()) {
+            throw new IllegalArgumentException();
+        }
+
+        for (final OrderTable savedOrderTable : savedOrderTables) {
+            if (!savedOrderTable.isEmpty() || Objects.nonNull(savedOrderTable.getTableGroupId())) {
+                throw new IllegalArgumentException();
+            }
+        }
+
+        final var tableGroup = new TableGroup(LocalDateTime.now(), savedOrderTables);
+
+        final TableGroup savedTableGroup = tableGroupDao.save(tableGroup);
+
+        final Long tableGroupId = savedTableGroup.getId();
+        for (final OrderTable savedOrderTable : savedOrderTables) {
+            savedOrderTable.setTableGroupId(tableGroupId);
+            savedOrderTable.setEmpty(false);
+            orderTableDao.save(savedOrderTable);
+        }
+        savedTableGroup.setOrderTables(savedOrderTables);
+
+        return savedTableGroup;
+    }
+
+    @Transactional
+    @Override
+    public void ungroup(final Long tableGroupId) {
+        final List<OrderTable> orderTables = orderTableDao.findAllByTableGroupId(tableGroupId);
+        validateTableGroupId(tableGroupId, orderTables);
+
+        final List<Long> orderTableIds = orderTables.stream()
+                .map(OrderTable::getId)
+                .collect(Collectors.toList());
+
+        if (orderDao.existsByOrderTableIdInAndOrderStatusIn(
+                orderTableIds, Arrays.asList(OrderStatus.COOKING.name(), OrderStatus.MEAL.name()))) {
+            throw new IllegalArgumentException();
+        }
+
+        for (final OrderTable orderTable : orderTables) {
+            orderTable.setTableGroupId(null);
+            orderTable.setEmpty(false);
+            orderTableDao.save(orderTable);
+        }
+    }
+
+    private void validateTableGroupId(final Long tableGroupId, final List<OrderTable> orderTables) {
+        if (orderTables.isEmpty()) {
+            throw new IllegalArgumentException("유효하지 않은 그룹 아이디 : " + tableGroupId);
+        }
+    }
+}
